Refactor: BoardGame to be a game model
